added two graph run
