+++ monitor.py
@@ -1,7 +1,5 @@
 #!/usr/bin/python3
 import argparse, socket, time, json, datetime, platform, psutil, requests, pprint, uuid
-from matplotlib import pyplot as plt
-from matplotlib.animation import FuncAnimation
 from os import system,name
 from tqdm import tqdm
 
@@ -194,31 +192,7 @@ def send_data(data):
 main()
 for i in tqdm(range(10), 'Data Gathering completed', colour='green'):
     time.sleep(.1)
-#graph
-try:
-    frame_len = 200
-    y = []
-    fig = plt.figure(figsize=(6,2))
 
-
-    def animation(i):
-        y.append(psutil.cpu_percent())
-        if len(y) <= frame_len:
-            plt.cla()
-            plt.plot(y, 'r', label='CPU usage(%)')
-        else:
-            plt.cla()
-            plt.plot(y[-frame_len:], 'r', label='cpu usage(%)')
-        plt.xlabel('Time (s)')
-        plt.ylabel('CPU usage(%)')
-        plt.title("REAL-TIME CPU Usage in percentage")
-        plt.legend(loc='upper right')
-
-
-    animate = FuncAnimation(plt.gcf(), animation, interval=1000)
-    plt.show()
-except KeyboardInterrupt:
-    print("Ctrl + C Entered. Exiting....")
-
+system('./gcpu.py & ./gnet.py')
 
 time.sleep(args.interval)
